Added a "Remove Bug" feature
Added a feature where the use can specify the bug ID and if it has been fixed then they can remove that bug from the database table.
It doesn't remove the bug from the list-box until the application is restarted which needs to be fixed/implemented. Have only been able to figure out how to clear all of the box or clear a certain line starting from 0 but this doesn't match up with the Bug ID number which is removed.